+++ src/Sylius/Bundle/CoreBundle/Resources/config/doctrine/model/PaymentMethod.orm.xml
@@ -27,6 +27,10 @@
                 </inverse-join-columns>
             </join-table>
         </many-to-many>
+        <many-to-one field="gatewayConfig" target-entity="Sylius\Bundle\PayumBundle\Model\GatewayConfig">
+            <cascade><cascade-all/></cascade>
+            <join-column name="gateway_config_id" referenced-column-name="id" nullable="true" on-delete="SET NULL" />
+        </many-to-one>
     </mapped-superclass>
 
 </doctrine-mapping>
