+++ src/Sylius/Bundle/CoreBundle/Resources/config/validation/PaymentMethod.xml
@@ -12,7 +12,7 @@
 -->
 
 <constraint-mapping xmlns="http://symfony.com/schema/dic/constraint-mapping" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/constraint-mapping http://symfony.com/schema/dic/services/constraint-mapping-1.0.xsd">
-    <class name="Sylius\Bundle\PayumBundle\Model\PaymentMethod">
+    <class name="Sylius\Component\Core\Model\PaymentMethod">
         <property name="gatewayConfig">
             <constraint name="Valid" />
         </property>
