--- src/Sylius/Bundle/PayumBundle/Resources/config/doctrine/model/PaymentMethod.orm.xml
@@ -1,22 +1,0 @@
-
-<!--
-
- This file is part of the Sylius package.
-
- (c) Paweł Jędrzejewski
-
- For the full copyright and license information, please view the LICENSE
- file that was distributed with this source code.
-
--->
-
-<doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping">
-
-    <mapped-superclass name="Sylius\Bundle\PayumBundle\Model\PaymentMethod" table="sylius_payment_method">
-        <many-to-one field="gatewayConfig" target-entity="Sylius\Bundle\PayumBundle\Model\GatewayConfig">
-            <cascade><cascade-all/></cascade>
-            <join-column name="gateway_config_id" referenced-column-name="id" nullable="true" on-delete="SET NULL" />
-        </many-to-one>
-    </mapped-superclass>
-
-</doctrine-mapping>
