+++ src/Sylius/Bundle/PayumBundle/spec/Action/Offline/ConvertPaymentActionSpec.php
@@ -17,7 +17,6 @@ use Payum\Core\Request\Capture;
 use Payum\Core\Request\Convert;
 use Payum\Offline\Constants;
 use PhpSpec\ObjectBehavior;
-use Prophecy\Argument;
 use Sylius\Bundle\PayumBundle\Action\Offline\ConvertPaymentAction;
 use Sylius\Component\Payment\Model\PaymentMethodInterface;
