+++ src/Sylius/Component/Core/Model/PaymentMethod.php
@@ -13,9 +13,11 @@ namespace Sylius\Component\Core\Model;
 
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
-use Sylius\Bundle\PayumBundle\Model\PaymentMethod as BasePaymentMethod;
+use Payum\Core\Model\GatewayConfigInterface;
 use Sylius\Component\Channel\Model\ChannelInterface as BaseChannelInterface;
+use Sylius\Component\Payment\Model\PaymentMethod as BasePaymentMethod;
 use Sylius\Component\Payment\Model\PaymentMethodTranslation;
+use Sylius\Component\Resource\Exception\UnsupportedMethodException;
 
 /**
  * @author Mateusz Zalewski <mateusz.zalewski@lakion.com>
@@ -26,6 +28,11 @@ class PaymentMethod extends BasePaymentMethod implements PaymentMethodInterface
      * @var Collection
      */
     protected $channels;
+
+    /**
+     * @var GatewayConfigInterface
+     */
+    protected $gatewayConfig;
 
     public function __construct()
     {
@@ -68,6 +75,42 @@ class PaymentMethod extends BasePaymentMethod implements PaymentMethodInterface
         if ($this->hasChannel($channel)) {
             $this->channels->removeElement($channel);
         }
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setGatewayConfig(GatewayConfigInterface $gatewayConfig)
+    {
+        $this->gatewayConfig = $gatewayConfig;
+    }
+
+    /**
+     * @return GatewayConfigInterface
+     */
+    public function getGatewayConfig()
+    {
+        return $this->gatewayConfig;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getGateway()
+    {
+        if (null !== $this->gatewayConfig) {
+            return $this->gatewayConfig->getGatewayName();
+        }
+
+        return null;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setGateway($gateway)
+    {
+        throw new UnsupportedMethodException('setGateway');
     }
 
     /**
