+++ src/Sylius/Component/Core/Model/PaymentMethodInterface.php
@@ -11,12 +11,22 @@
 
 namespace Sylius\Component\Core\Model;
 
-use Sylius\Bundle\PayumBundle\Model\PaymentMethodInterface as BasePaymentMethodInterface;
+use Payum\Core\Model\GatewayConfigInterface;
 use Sylius\Component\Channel\Model\ChannelsAwareInterface;
+use Sylius\Component\Payment\Model\PaymentMethodInterface as BasePaymentMethodInterface;
 
 /**
  * @author Mateusz Zalewski <mateusz.zalewski@lakion.com>
  */
 interface PaymentMethodInterface extends BasePaymentMethodInterface, ChannelsAwareInterface
 {
+    /**
+     * @param GatewayConfigInterface $gateway
+     */
+    public function setGatewayConfig(GatewayConfigInterface $gateway);
+
+    /**
+     * @return GatewayConfigInterface
+     */
+    public function getGatewayConfig();
 }
