+++ src/Sylius/Component/Core/composer.json
@@ -23,6 +23,7 @@
         "php": "^5.6|^7.0",
 
         "knplabs/gaufrette": "^0.2",
+        "payum/payum": "^1.3",
         "sylius/addressing": "^1.0",
         "sylius/attribute": "^1.0",
         "sylius/channel": "^1.0",
