+++ src/Sylius/Component/Core/spec/Model/PaymentMethodSpec.php
@@ -13,11 +13,13 @@ namespace spec\Sylius\Component\Core\Model;
 
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
+use Payum\Core\Model\GatewayConfigInterface;
 use PhpSpec\ObjectBehavior;
 use Sylius\Component\Core\Model\ChannelInterface;
 use Sylius\Component\Core\Model\PaymentMethod;
 use Sylius\Component\Core\Model\PaymentMethodInterface;
-use Sylius\Bundle\PayumBundle\Model\PaymentMethod as BasePaymentMethod;
+use Sylius\Component\Payment\Model\PaymentMethod as BasePaymentMethod;
+use Sylius\Component\Resource\Exception\UnsupportedMethodException;
 
 /**
  * @author Mateusz Zalewski <mateusz.zalewski@lakion.com>
@@ -54,6 +56,27 @@ final class PaymentMethodSpec extends ObjectBehavior
 
         $this->removeChannel($channel);
         $this->hasChannel($channel)->shouldReturn(false);
+    }
+
+    function its_gateway_is_gateway_config_object(GatewayConfigInterface $gatewayConfig)
+    {
+        $this->setGatewayConfig($gatewayConfig);
+        $this->getGatewayConfig()->shouldReturn($gatewayConfig);
+    }
+
+    function it_does_not_support_set_gateway_method()
+    {
+        $this->shouldThrow(UnsupportedMethodException::class)->during('setGateway', ['gateway']);
+    }
+
+    function it_returns_name_from_gateway_config_while_getting_a_gateway(GatewayConfigInterface $gatewayConfig)
+    {
+        $this->getGateway()->shouldReturn(null);
+
+        $gatewayConfig->getGatewayName()->willReturn('gateway');
+        $this->setGatewayConfig($gatewayConfig);
+
+        $this->getGateway()->shouldReturn('gateway');
     }
 
     public function getMatchers()
